Add Tests for Technical Indicators
Added testthat infrastructure.

Created data file for testing.

Moved quantmod to depends. This was needed so quantmod, xts and TTR are loaded when testing. The data file is an xts object so xts needs to be loaded before tests are run.

Created tests for stochRSI: references issues #3 and #5
created tests for envelope: references issues  #4 and #6