Fixed issues with travis and r cmd check.
Moved quantmod from depends to imports. There were some loose calls in the code that required quantmod to be loaded for the functions to work correctly. Fixed this by referring all functions from other packages with package::function.

Removed Matrix::rowMeans for base rowMeans. Otherwise Matrix would need to be in imports.

Removed setNames for the same reason. This is from stats and using the names function does the trick as well.

Added license file to make sure R cmd check is not complaining about this.

Fixed 2 mistakes in the functions ATR_stop and heikin_ashi. This also fixed the weird workaround I had in the test files for loading the data in the global environment instead of just the test environment. The tests now function as expected.

Fixed issue in envelope function that needed quantmod as a dependend package. Now using TTR's EMA and SMA functions instead of a do.call.

Commit references issues #11, #17, #18