Adds alert marker and market thermometer
Created alert marker code in alert_marker.R file.  Also documentation has been created and used the @family roxygen to link the documentation to the other Elder functions.

Created market thermometer code in the market_thermometer.R file. Documentation has been added and linked with the @family.

Added @family to the documentation in the safezone stop.

Added @family to the documentation of all the protective stops.

Namespace has been updated to reflect the addition of the two functions.